deleted admin folder and less file, imported and rendered table to new employee component
