+++ src/components/common/Table/Checkbox.js
@@ -1,0 +1,28 @@
+import React, { useState } from 'react';
+
+const CheckboxComponent = props => {
+  const [selected, setSelected] = useState([]);
+  const selectedRowKeys = selected;
+
+  const onSelectChange = selectedRowKeys => {
+    setSelected(selectedRowKeys);
+    const selectedUsers = [];
+    selectedRowKeys.map(key => {
+      props.map(user => {
+        if (user.key === key) {
+          selectedUsers.push(user);
+        }
+      });
+    });
+    console.log(`Selected User Data:`, selectedUsers);
+  };
+
+  const rowSelection = {
+    selectedRowKeys,
+    onChange: onSelectChange,
+  };
+
+  return rowSelection;
+};
+
+export default CheckboxComponent;
