--- src/components/pages/AdminPages/Employees.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import { TableComponent } from '../../common/index';
-
-function Employees() {
-  return (
-    <>
-      <TableComponent />
-    </>
-  );
-}
-
-export default Employees;
