--- src/components/pages/AdminPages/index.js
@@ -1,1 +1,0 @@
-export { default as EmployeesPage } from './Employees';
