--- src/styles/components/_table.less
@@ -1,30 +1,0 @@
-.ant-table-wrapper {
-  width: 95%;
-  height: 100%;
-  margin-left: 2.5%;
-  padding: 1px;
-  background: #f0f0f0;
-  border: 1px solid #f0f0f0;
-  border-radius: 2px;
-
-  .ant-tag {
-    margin-bottom: 7px;
-  }
-
-  #components-dropdown-demo-arrow .ant-btn {
-    margin-right: 8px;
-    margin-bottom: 8px;
-  }
-
-  .ant-row-rtl #components-dropdown-demo-arrow .ant-btn {
-    margin-right: 0;
-    margin-bottom: 8px;
-    margin-left: 8px;
-  }
-
-  .editable-row .ant-form-item-explain {
-    position: absolute;
-    top: 100%;
-    font-size: 12px;
-  }
-}
